Fix map method & pokemons as a module
